Rename scan_path_algorithm.py to algorithm.py
