+++ pupil_src/shared_modules/scan_path/algorithm.py
